requirements.txt
Flask==2.3.3
yt-dlp==2023.7.6
python-vlc==3.0.18121